Change links to .html
Moving to static html pages will force the links to
require the .html suffix. I'm sure there is some way to
configure this away, and we may want to look into it.

If we are concerned about previously shared links to
pages other than the home page, we might not find
this acceptable.